Add first version of database connection
The cloud database used is Microsoft Azure SQL Database. The
corresponding documentation is here:
https://docs.microsoft.com/en-us/azure/sql-database/

For the moment, the database firewall only accepts one IP-address.
Every other connections are aborted.

The connection permits to insert, search and delete one element in a
simple “Example” table containing two columns : an ID (int) and a NAME
(varchar(20)). There is no automated test but a simple test class. If
we play with it concurrently with another database software (in my case
Visual Studio Code), we can check to behaviour of the DatabaseConnector
class.