The Server Optimal protocol is done and tested
Note on ExecutorService class:
If a thread throws an error, it does not mean that a thread started
before another or that a pool of threads runs out of memory. It means
that a thread encounters a problem at runtime and throws an
exception. The ExecutorService class does not tell you which exception
was thrown, it just tells which thread throws it. So, the problem does
not come from parallelism.

Solution of the problem:
When the client sends the response of the challenge, it does not inform
the server which protocol is used. (Since sending a response is the
same for each protocol.) The error comes from the fact that the
server’s first action is to get the protocol used.
So, sending the protocol used even when the client sends a response
permits to correct the problem.